feat(trails): declare native bun release binding (#714)
## Context

Part of the Release Rules M3 follow-through for TRL-938.

This branch declares the Trails-owned Bun release path as a native release binding instead of leaving publish and registry behavior in root scripts.

## What Changed

- Added release binding descriptors for native and adapter release bindings.
- Promoted the Bun publish and registry preflight implementation into `@ontrails/trails/release`.
- Kept the existing root scripts as thin compatibility wrappers around the package-owned implementation.
- Added focused release binding coverage and a branch-local changeset for `@ontrails/trails`.

## Verification

- Focused release binding, release smoke, MCP, publish, and registry tests passed.
- `bun run check`
- `bun run test`
- `bun run build`
- `bun run wayfinder:dogfood`
- `bun run dogfood:packed`
- `bun run publish:check`
- `gt submit --stack --draft --restack --no-edit --no-interactive --dry-run`
- Real Graphite submit pre-push passed: dead-code, format, lint, ast-grep, test, typecheck, Warden.

Known Warden warnings are the existing internal `wayfind.*` reachability warnings.
