+++ .changeset/trl-938-native-bun-release-binding.md
@@ -1,0 +1,5 @@
+---
+"@ontrails/trails": patch
+---
+
+Expose the native Bun release binding from `@ontrails/trails/release` and keep publish and registry scripts as compatibility wrappers.
